// cleint/src/models/INftData.ts
export interface INFTDATA {
    idnft: number;
    iduser: number;
    name: string;
    description: string;
    ethername: string;
    originals: string;
    image: string ;
    price: number;
    highest: number;
    username: string;
    email: string;
    password: string;
    bio: string;
    discordlink: string;
    website: string;
    youtube: string;
    twitter: string;
    instagram: string;
    facebook: string;
    token: string | null;
    avatar: string;
}